requirements.txt
streamlit>=1.31.0
langchain>=0.1.0
langchain-community>=0.0.25
langchain-core>=0.1.0
langgraph>=0.0.29
langgraph-prebuilt>=0.1.0
langgraph-checkpoint>=1.0.0
langchain-openai>=0.0.5
langchain-text-splitters>=0.0.1
chromadb>=0.4.22
dashscope>=1.14.0
pypdf>=4.0.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
tiktoken>=0.5.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
